Change namespace to PuppetX
Upon further consideration, I should follow the newer conventions with
custom ruby extentions. Should this code ever make its way into core, it
will be in the proper 'Puppet' namespace. Until that day, 'PuppetX' is
where this will live.

This commit converts the 'Puppet' namespace to 'PuppetX' in the code.
This commit also includes 'PuppetX' being added to the README.